+++ demo-shell-ng2/app/app.component.html
@@ -21,6 +21,7 @@
                 <a class="mdl-navigation__link" data-automation-id="files" href="" routerLink="/files">DocumentList</a>
                 <a class="mdl-navigation__link" data-automation-id="activiti" href="" routerLink="/activiti">Process Services</a>
                 <a class="mdl-navigation__link" data-automation-id="form" href="" routerLink="/form">Form</a>
+                <a class="mdl-navigation__link" data-automation-id="form" href="" routerLink="/form-list">Form List</a>
                 <a class="mdl-navigation__link" data-automation-id="login" href="" routerLink="/login">Login</a>
                 <a class="mdl-navigation__link" data-automation-id="settings" href="" routerLink="/settings">Settings</a>
             </nav>
