+++ demo-shell-ng2/app/components/form/form-demo.component.css
@@ -1,8 +1,3 @@
 .form-container {
     padding: 10px;
 }
-
-.store-form-container{
-    width: 80%;
-    height: 80%;
-}
